requirements.txt
PySide6==6.5.0
PySimpleGUIQt6==0.36.0
yeelight==0.7.10